+++ visual-testing-app/src/components/custom-form-modal-page/custom-form-modal-page.visualroute.tsx
@@ -1,4 +1,5 @@
 import { Formik } from 'formik';
+import { useHistory } from 'react-router-dom';
 import {
   SearchIcon,
   FlameIcon,
@@ -7,23 +8,18 @@ import {
 import TextField from '@commercetools-uikit/text-field';
 import IconButton from '@commercetools-uikit/icon-button';
 import { CustomFormModalPage } from '@commercetools-frontend/application-components';
-import { Suite, Spec } from '../../test-utils';
+import { NestedPages, Suite } from '../../test-utils';
 
 export const routePath = '/custom-form-modal-page';
 
-type ContainerProps = {
-  portalId: string;
-} & Partial<Parameters<typeof CustomFormModalPage>[0]>;
+type ContainerProps = Partial<Parameters<typeof CustomFormModalPage>[0]>;
 type FormValues = {
   email: string;
 };
 
-const ModalPageWithPortalParentSelector = ({
-  portalId,
-  ...props
-}: ContainerProps) => (
-  <>
-    <div id={portalId} style={{ position: 'relative', height: '750px' }} />
+const ModalPageWithPortalParentSelector = (props: ContainerProps) => {
+  const history = useHistory();
+  return (
     <Formik<FormValues>
       initialValues={{ email: '' }}
       onSubmit={() => undefined}
@@ -31,18 +27,15 @@ const ModalPageWithPortalParentSelector = ({
       {(formikProps) => (
         <CustomFormModalPage
           title="Lorem ipsum"
-          isOpen={true}
-          onClose={() => undefined}
+          isOpen
+          onClose={() => history.push(routePath)}
           subtitle="Lorem ipsum dolor sit amet, consectetur adipiscing elit."
-          getParentSelector={() =>
-            document.querySelector(`#${portalId}`) as HTMLElement
-          }
           {...props}
         >
           <TextField
             name="email"
             title="Email"
-            isRequired={true}
+            isRequired
             value={formikProps.values.email}
             errors={
               TextField.toFieldErrors<FormValues>(formikProps.errors).email
@@ -55,71 +48,89 @@ const ModalPageWithPortalParentSelector = ({
         </CustomFormModalPage>
       )}
     </Formik>
-  </>
-);
+  );
+};
 ModalPageWithPortalParentSelector.displayName =
   'ModalPageWithPortalParentSelector';
 
 export const Component = () => (
   <Suite>
-    <Spec label="CustomFormModalPage" size="xl">
-      <ModalPageWithPortalParentSelector portalId="custom-form-modal-page-default"></ModalPageWithPortalParentSelector>
-    </Spec>
-    <Spec
-      label="CustomFormModalPage - with the static exposed form controls"
-      size="xl"
-    >
-      <ModalPageWithPortalParentSelector
-        formControls={
-          <>
-            <CustomFormModalPage.FormSecondaryButton
-              onClick={() => undefined}
+    <NestedPages
+      title="Custom form modal pages"
+      basePath={routePath}
+      pages={[
+        {
+          path: 'custom-form-modal-page-default',
+          spec: <ModalPageWithPortalParentSelector />,
+        },
+        {
+          path: 'custom-form-modal-page-default-controls',
+          spec: (
+            <ModalPageWithPortalParentSelector
+              formControls={
+                <>
+                  <CustomFormModalPage.FormSecondaryButton
+                    onClick={() => undefined}
+                  />
+                  <CustomFormModalPage.FormPrimaryButton
+                    onClick={() => undefined}
+                  />
+                  <CustomFormModalPage.FormDeleteButton
+                    onClick={() => undefined}
+                  />
+                </>
+              }
             />
-            <CustomFormModalPage.FormPrimaryButton onClick={() => undefined} />
-            <CustomFormModalPage.FormDeleteButton onClick={() => undefined} />
-          </>
-        }
-        portalId="custom-form-modal-page-default-controls"
-      />
-    </Spec>
-    <Spec label="CustomFormModalPage - with other custom controls" size="xl">
-      <ModalPageWithPortalParentSelector
-        formControls={
-          <>
-            <IconButton
-              label="Search"
-              icon={<SearchIcon />}
-              onClick={() => undefined}
+          ),
+        },
+        {
+          path: 'custom-form-modal-page-custom-controls',
+          spec: (
+            <ModalPageWithPortalParentSelector
+              formControls={
+                <>
+                  <IconButton
+                    label="Search"
+                    icon={<SearchIcon />}
+                    onClick={() => undefined}
+                  />
+                  <IconButton
+                    label="Update"
+                    icon={<FlameIcon />}
+                    onClick={() => undefined}
+                  />
+                  <IconButton
+                    label="Delete"
+                    icon={<BinLinearIcon />}
+                    onClick={() => undefined}
+                  />
+                </>
+              }
             />
-            <IconButton
-              label="Update"
-              icon={<FlameIcon />}
-              onClick={() => undefined}
+          ),
+        },
+        {
+          path: 'custom-form-modal-page-hidden-controls',
+          spec: (
+            <ModalPageWithPortalParentSelector
+              formControls={
+                <>
+                  <CustomFormModalPage.FormSecondaryButton
+                    onClick={() => undefined}
+                  />
+                  <CustomFormModalPage.FormPrimaryButton
+                    onClick={() => undefined}
+                  />
+                  <CustomFormModalPage.FormDeleteButton
+                    onClick={() => undefined}
+                  />
+                </>
+              }
+              hideControls
             />
-            <IconButton
-              label="Delete"
-              icon={<BinLinearIcon />}
-              onClick={() => undefined}
-            />
-          </>
-        }
-        portalId="custom-form-modal-page-custom-controls"
-      ></ModalPageWithPortalParentSelector>
-    </Spec>
-    <Spec label="CustomFormModalPage - with hidden controls" size="xl">
-      <ModalPageWithPortalParentSelector
-        formControls={
-          <>
-            <CustomFormModalPage.FormSecondaryButton
-              onClick={() => undefined}
-            />
-            <CustomFormModalPage.FormPrimaryButton onClick={() => undefined} />
-            <CustomFormModalPage.FormDeleteButton onClick={() => undefined} />
-          </>
-        }
-        hideControls={true}
-        portalId="custom-form-modal-page-hidden-controls"
-      ></ModalPageWithPortalParentSelector>
-    </Spec>
+          ),
+        },
+      ]}
+    />
   </Suite>
 );