+++ visual-testing-app/src/components/custom-form-modal-page/custom-form-modal-page.visualspec.ts
@@ -1,13 +1,19 @@
 import percySnapshot from '@percy/puppeteer';
 import { HOST } from '../../constants';
 
-describe(`CustomFormModalPage`, () => {
+describe.each`
+  path
+  ${'custom-form-modal-page-default'}
+  ${'custom-form-modal-page-default-controls'}
+  ${'custom-form-modal-page-custom-controls'}
+  ${'custom-form-modal-page-hidden-controls'}
+`('CustomFormModalPage $path', ({ path }) => {
   beforeAll(async () => {
-    await page.goto(`${HOST}/custom-form-modal-page`);
+    await page.goto(`${HOST}/custom-form-modal-page/${path}`);
   });
 
-  it('Default', async () => {
-    await expect(page).toMatch('CustomFormModalPage');
-    await percySnapshot(page, `CustomFormModalPage`);
+  it('renders page', async () => {
+    await expect(page).toMatch('Lorem ipsum');
+    await percySnapshot(page, `CustomFormModalPage/${path}`);
   });
 });