+++ packages/application-components/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @commercetools-frontend/application-components
 
+## 21.5.1
+
+### Patch Changes
+
+- [#2588](https://github.com/commercetools/merchant-center-application-kit/pull/2588) [`787692ac`](https://github.com/commercetools/merchant-center-application-kit/commit/787692accfed47c5dc50ef9b711e38d76ae97a87) Thanks [@emmenko](https://github.com/emmenko)! - Leftover of https://github.com/commercetools/merchant-center-application-kit/pull/2586
+
+## 21.5.0
+
+### Patch Changes
+
+- [#2586](https://github.com/commercetools/merchant-center-application-kit/pull/2586) [`2f58b2d6`](https://github.com/commercetools/merchant-center-application-kit/commit/2f58b2d619fbc9d30e7d89604aba97f0c0ac9325) Thanks [@emmenko](https://github.com/emmenko)! - Fix type declarations for `<PortalsContainer>` as `defaultProps` and `forwardRef` do not work well together.
+
 ## 21.4.0
 
 ### Minor Changes